backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
autoflake==2.3.1
blessed==1.20.0
coverage==7.6.1
Django==5.1.1
django-cors-headers==4.4.0
django-picklefield==3.2
django-q2==1.7.2
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
importlib_metadata==8.5.0
inflection==0.5.1
isort==5.13.2
packaging==24.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==3.5.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
types-python-dateutil==2.9.0.20240906
uritemplate==4.1.1
wcwidth==0.2.13
yapf==0.40.2
zipp==3.20.2
